requirements.txt
streamlit
pandas
Pillow
joblib
xgboost==2.1.2
scikit-learn
imbalanced-learn
